ReAct_Agents/requirements.txt
accelerate==0.23.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
ale-py==0.8.1
altair==5.0.1
anyio==3.7.1
argcomplete==2.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
Babel==2.12.1
backcall==0.2.0
backgroundremover==0.2.4
backoff==2.2.1
beautifulsoup4==4.12.2
black==23.7.0
bleach==6.0.0
blis==0.7.11
# box2d-py==2.3.5
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.4.5
click==8.1.6
cloudpathlib==0.15.1
cloudpickle==2.2.1
cmake==3.27.2
coloredlogs==15.0.1
comm==0.1.3
commandlines==0.4.1
confection==0.1.3
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
cymem==2.0.8
dataclasses-json==0.5.14
datasets==2.14.5
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
diffusers==0.21.2
dill==0.3.7
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.2
executing==1.2.0
Faker==19.12.0
Farama-Notifications==0.0.4
fastapi==0.99.1
fastjsonschema==2.18.0
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.42.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
genson==1.2.2
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.23.3
google-auth-httplib2==0.1.0
google-search-results==2.4.2
googleapis-common-protos==1.60.0
gradio==3.40.1
gradio_client==0.4.0
greenlet==2.0.2
gymnasium==0.29.1
h11==0.14.0
hsh==1.1.0
httpcore==0.17.3
httplib2==0.22.0
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib-resources==6.0.1
inflect==5.6.2
install==1.3.5
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
isoduration==20.11.0
isort==5.12.0
jaraco.context==4.3.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.4
langchain==0.0.310
langchain-experimental==0.0.32
langchainhub==0.1.13
langcodes==3.3.0
langsmith==0.0.43
lark==1.1.7
lazy_loader==0.3
linkify-it-py==2.0.2
lit==16.0.6
# llama-index==0.8.51
llvmlite==0.40.1
loguru==0.7.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.1
monotonic==1.6
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
notebook==7.0.2
notebook_shim==0.2.3
numba==0.57.1
numexpr==2.8.5
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
opencv-python==4.8.0.76
orjson==3.9.4
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
path==16.7.1
pathspec==0.11.2
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
pil-plus==0.2
Pillow==10.0.0
pinecone-client==2.2.4
platformdirs==3.5.1
posthog==3.0.1
prance==0.22.2.22.0
preshed==3.0.9
proglog==0.1.10
prometheus-client==0.17.1
prompt-toolkit==3.0.38
protobuf==4.24.0
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pydub==0.25.1
pygame==2.5.2
Pygments==2.15.1
PyMatting==1.1.8
pyparsing==3.0.9
pypdf==3.15.0
PyPika==0.48.9
pyrsistent==0.16.1
PySnooper==1.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safetensors==0.3.2
scikit-image==0.21.0
scipy==1.11.1
semantic-version==2.10.0
Send2Trash==1.8.2
Shimmy==0.2.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.19
srsly==2.4.8
stack-data==0.6.2
starlette==0.27.0
swig==4.1.1
sympy==1.12
tenacity==8.2.2
terminado==0.17.1
text-generation==0.6.0
thinc==8.2.1
tifffile==2023.8.12
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.31.0
triton==2.0.0
typed-ast==1.5.5
typer==0.9.0
types-requests==2.31.0.10
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.7
# urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.17.0
validators==0.21.0
waitress==2.1.2
wasabi==1.1.2
watchfiles==0.19.0
wcwidth==0.2.6
weasel==0.3.2
weaviate-client==3.22.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
widgetsnbextension==3.5.2
wikipedia==1.4.0
wolframalpha==5.0.0
xmltodict==0.13.0
xxhash==3.3.0
yarl==1.9.2
zipp==3.17.0
